refactor(hidtransport): complete Windows HID binding (HidD + SetupDi + R/W)
hidtransport previously exposed only SetFeature on the device
side, and used the Raw Input path (user32) for discovery. That
left a large gap between 'the only thing Simagic needs today' and
'what a HID device on Windows actually exposes'.

Rewritten to be a complete Go binding to the Windows HID API,
matching the public HidD_* surface plus SetupDi_* and the file
handles' ReadFile/WriteFile. The Scanner path is now the canonical
Microsoft-documented 'Finding and Opening a HID Collection'
(SetupDiGetClassDevs / EnumDeviceInterfaces /
GetDeviceInterfaceDetail) instead of mixing Raw Input enumeration
with HIDClass device handles.

New surface on *Transport:

  Attributes / ManufacturerString / ProductString /
  SerialNumberString / IndexedString / PhysicalDescriptor
  PreparsedData + (*PreparsedData).Capabilities
  SetFeature / GetFeature / SetOutputReport / GetInputReport
  Read / Write   (interrupt IN / OUT via ReadFile / WriteFile)
  FlushQueue / NumInputBuffers / SetNumInputBuffers

New surface on Scanner: same Scan() signature, but the descriptors
now carry SerialNumber and the correct Usage/UsagePage (lifted from
the preparsed data rather than from RAWINPUTDEVICEINFO).

Type changes:

  DeviceDescriptor.VersionNumber: uint16 (was uint32) to match
  HIDD_ATTRIBUTES; hpr.DeviceInfo.VersionNumber follows.
  hpr.DeviceInfo.VersionNumber: uint16 too.

HidP_GetCaps returns NTSTATUS (HIDP_STATUS_SUCCESS = 0x00110000),
not a BOOL; the Scanner checks that NTSTATUS explicitly instead
of trusting the syscall return.

No API change to hpr or to any driver. Simagic still uses
Open + SetFeature; the underlying transport is now the complete
HID surface for whoever needs it next.

+++ pkg/hpr/types.go
@@ -94,7 +94,7 @@ type DeviceInfo struct {
 
 	VendorID      uint16
 	ProductID     uint16
-	VersionNumber uint32
+	VersionNumber uint16
 
 	UsagePage uint16
 	Usage     uint16
